WP 560- The Effects of Schooling on Costless Health Maintenance: Overweight Adolescents and Children in Rural China
Obesity is an important global health problem. Although obesity is not directly related to access to health care or constrained by resource deprivation, overweight status is predominantly found in poor, less-educated populations. This paper seeks to identify the causal role of schooling in affecting obesity among children and adolescents, using new estimation methods that exploit unique panel data on young twins in China. The estimates indicate that higher levels of schooling negatively affect being overweight and positively affect healthy behavior, with a large component of the causal effects due to increased information on the benefits of maintaining a healthy weight.