WP 559- Cashing In (and Out): Experimental Evidence on the Effects of Mobile Money in Malawi
Mobile money has spread rapidly across Africa since it was first introduced in Kenya in 2007, and has been extensively studied in recent years. However, identification is challenging because mobile money is typically rapidly adopted, making it difficult to preserve a control group. In this paper, we add to previous literature with an RCT conducted among microentrepreneurs in urban Malawi in 2017-18, where usage of mobile money was still modest, even though a nascent mobile money agent network existed. Treatment was three-pronged: assistance in opening a mobile money account, training on how to perform basic transactions, and a withdrawal fee waiver. We find that the majority of people opened accounts and used them extensively. We find strong evidence that treated respondents reallocated labor from business to agriculture, and we find more mixed evidence of an increase in expenditures. In contrast to the existing literature, effects appear to be driven by using the accounts to save rather than to make transfers.