WP 553- Demand Estimation with Strategic Complementarities: Sanitation in Bangladesh
For many products, the utility of adoption depends on the share of other households that adopt. We estimate a structural model of demand that allows for these interdependencies. We apply our model to the adoption of household latrines – a technology that has large consequences for public health. We estimate the model using data from a large-scale experiment covering over 18,000 households in 380 communities in rural Bangladesh, where we randomly assigned incentives to purchase latrines. Subsidies were randomly assigned at the household level to identify the direct effect of price, and subsidy saturation was randomly varied at the community level to identify strategic complementarities in demand. We conduct counter-factual simulations to analyze the policymaker’s tradeoffs along price, saturation and scope margins: To raise aggregate latrine adoption, is it better to intensely subsidize a few, or widely disperse subsidies across households or communities? We also analyze the effects of targeting subsidies on the basis of household poverty, social position, or neighborhood population density. Finally, we use additional experiments to explore mechanisms underlying the complementarity in demand, and find that shame and changing social norms are driving factors.