WP 546- Identifying the Costs of a Public Health Success: Arsenic Well Water Contamination and Productivity in Bangladesh
In the 1970s and 1980s the government of Bangladesh, with the support and financing of the United Nations Children’s Fund, promoted the digging of tube wells to provide clean drinking water and reduce the incidence of diarrheal disease. Prior to this, drinking water came from surface water sources, which were identified as a principal source of diarrheal disease. In the late 1990s, however, evidence indicated that groundwater, unlike surface water and by then the main source of water for drinking, irrigation and cooking in Bangladesh, was contaminated by naturally-occurring arsenic in 59 of the country’s 64 districts. While there is now a large body of evidence that the consumption of arsenic-contaminated water is manifested in higher levels of morbidity and new symptoms (e.g., skin lesions), there is little evidence on the economic consequences of arsenic ingestion and retention in the body in Bangladesh. In this paper, we provide new estimates of the effects of the ingestion and retention of inorganic arsenic
on direct measures of cognitive and physical capabilities as well as on the schooling attainment, occupational structure, entrepreneurship and incomes of the rural Bangladesh population. We also provide new estimates of the effects of the consumption of foods grown and cooked in arsenic-contaminated water on individual arsenic concentrations.