WP 545- Understanding Human Trafficking Using Victim-Level Data
Quantitative research on human trafficking is scant due to lack of data. This study makes use of a unique survey we collected on former victims of trafficking and vulnerable women and girls in the Philippines. We start by exploring the correlates of trafficking and show that household composition (in particular the presence of older sisters) and plausibly exogenous measures of health and economic shocks predict the likelihood of being trafficked. We then study the effects of trafficking on victims’ intertemporal and risk preferences using entropy balancing. We find that trafficking victims are not differentially patient, but they are more risk-loving. Our novel data and findings are pertinent to the design of policies intending to prevent trafficking and reintegrate victims.