WP 544- Worst-Case Bounds on R&D and Pricing Distortions: Theory and Disturbing Conclusions if Consumer Values Follow the World Income Distribution
We prove that, for general demand and cost conditions and market structures, the fraction of first-best surplus that a monopolist is unable to extract in a market provides a tight upper bound on the relative distortions arising from firms’ equilibrium decisions at all margins (entry and pricing). Continuing with this worst-case perspective, we show that a symmetrically truncated Zipf (STRZ) distribution of consumer values generates the lowest producer surplus among those with a given mean and maximum value. This allows us to relate potential deadweight loss from all margins in a market to the Zipf-similarity of its demand curve. The STRZ distribution also bounds deadweight loss at just the pricing margin. We leverage existing results from industrial organization (e.g., on demand curvature) and statistics (e.g., on the relation between means and medians) to bound producer surplus in an array of important special cases. Calibrations based on the world distribution of income generate extremely Zipf-similar demand curves, with disturbing consequences for potential deadweight loss in global markets. We gauge the extent to which various policies—such as progressive taxation or price discrimination—can ameliorate potential deadweight loss.