WP 540- Bride Price and Female Education
Although it is well known that traditional cultural practices can play an important role in development, we still have little understanding of what this means for development policy. To improve our understanding of this issue, we examine how the effects of school construction on girls’ education vary with a widely-practiced marriage custom called bride price, which is a payment made by the husband and/or his family to the wife’s parents at marriage. We begin by developing a model of educational choice with and without bride price. The model generates a number of predictions that we test in two countries that have had large-scale school construction projects, Indonesia and Zambia. Consistent with the model, we find that for groups that practice the custom of bride price, the value of bride price payments that the parents receive tend to increase with their daughter’s education. As a consequence, the probability of a girl being educated is higher among bride price groups. The model also predicts that families from bride price groups will be the most responsive to policies, like school construction, that are aimed at increasing female education. Studying the INPRES school construction program in Indonesia, as well as a similar program in Zambia, we find evidence consistent with this prediction. Although the program had no discernible effect on the education of girls from groups without bride price, it had large positive effects for girls from groups with a bride price. The findings emphasize the importance of the marriage market as a driver of educational investment and provide an example of how the cultural context of a society can be crucial for the effectiveness of development policy.