WP 539- The Effects of Fair Trade Certification: Evidence From Coffee Producers in Costa Rica
We examine the effects of Fair Trade (FT) certification of coffee on producers and households in Costa Rica. Examining the production dynamics of the universe of Costa Rican coffee mills from 1999–2014, we find that FT certification is associated with a higher sales price, greater sales, and more revenues. As expected, these effects are greater when global coffee prices are lower and the FT guaranteed minimum price is binding. Looking at households, we find evidence that FT is associated with higher incomes for all families, but especially for those working in the coffee sector. However, we also find that, within this sector, the benefits are not evenly distributed. Skilled coffee growers benefit from FT, intermediaries are hurt, and unskilled workers are unaffected. Thus, although FT creates sizable benefits (on average), it also results in a redistribution from intermediaries to farmers. Lastly, we also find evidence of positive effects of FT certification on the education of high-school-aged children, which is most likely due to the presence of scholarship programs that are funded by FT premiums.