WP 538- Social Structure and Conflict: Evidence from Sub-Saharan Africa
We test the long-standing hypothesis that ethnic groups that are organized around ‘segmentary lineages’ are more prone to conflict and civil war. Ethnographic accounts suggest that in segmentary lineage societies, which are characterized by strong allegiances to distant relatives, individuals are obligated to come to the defense of fellow lineage members when they become involved in conflicts. As a consequence, small disagreements often escalate to larger-scale conflicts involving many individuals. We test for this link between segmentary lineage and conflict across 145 African ethnic groups in sub-Saharan Africa. Using a number of estimation strategies, including an RD design at ethnic boundaries, we find that segmentary lineage societies experience more conflicts and ones that are longer in duration and larger in scale. We also find that the previously-documented relationship between adverse rainfall shocks and conflict within Africa is only found within segmentary lineage societies.