WP 533- Unpacking a Multi-Faceted Program to Build Sustainable Income for the Very Poor
A multi-faceted program comprising a grant of productive assets, training, coaching, and savings has been found to build sustainable income for those in extreme poverty. We focus on two important questions: whether a mere grant of productive assets would generate similar impacts (it does not), and whether access to a savings account and a deposit collection service would generate similar impacts (it does not).