WP 529- Environmental Externalities and Free-Riding in the Household
Water use and electricity use, which generate negative environmental externalities, are susceptible to a second externality problem: with household-level billing, each person enjoys private benefits of consumption but shares the cost with other household members. If individual usage is imperfectly observed (as is typical for water and electricity) and family members are imperfectly altruistic toward one another, households overconsume even from their own perspective. We develop this argument and test its prediction that intrahousehold free-riding dampens price sensitivity. We do so in the context of water use in urban Zambia by combining billing records, randomized price variation, and a lab-experimental measure of intrahousehold altruism. We find that more altruistic households are considerably more price sensitive than are less altruistic households. Our results imply that the socially optimal price needs to be set to correct both the environmental externality and also the intrahousehold externality.