WP 528- Goals and Gaps: Educational Careers of Immigrant Children
We study the educational choices of children of immigrants in a tracked school system. We first show that immigrant boys in Italy enroll disproportionately into vocational high schools, as opposed to technical and academically-oriented high schools, compared to natives of similar ability. Immigrant girls, instead, choose similar schools as native ones. We then estimate the impact of a large-scale, randomized intervention providing tutoring and career counseling to high-ability immigrant students. Male treated students increase their probability of enrolling into the high track to the same level of natives, also closing the gap in terms of grade retention. There are no significant effects on immigrant females, who exhibit similar choices and performance as native ones in absence of the intervention. Increases in academic motivation and the resulting changes in teachers’ recommendation regarding high school choice explain a sizable portion of the effect, while the effect of increases in cognitive skills is negligible. Finally, we find positive spillovers on immigrant classmates of treated students, while there is no effect on native classmates.