WP 527- The Long-run Effects of Agricultural Productivity on Conflict, 1400-1900
This paper provides evidence of the long-run effects of a permanent increase in agricultural productivity on conflict. We construct a newly digitized and geo-referenced dataset of battles in Europe, the Near East and North Africa covering the period between 1400 and 1900 CE. For variation in permanent improvements in agricultural productivity, we exploit the introduction of potatoes from the Americas to the Old World after the Columbian Exchange. We find that the introduction of potatoes permanently reduced conflict for roughly two centuries. The results are driven by a reduction in civil conflicts.