WP 525- Vertical Integration, Supplier Behavior, and Quality Upgrading among Exporters
We study the relationship between firms’ output quality and their choice of organizational structure. To do so, we use data on each step of the production and transaction chain that makes up Peruvian fishmeal manufacturing. We first show that quality upgrading is an important motive for vertically integrating. Firms integrate suppliers when the quality premium—the relative price of high quality output—rises for exogenous reasons, but not when average or low quality prices rise. The greater a firm’s scope for shifting low to high quality production, the greater its integration response. We then show that integration changes suppliers’ production behavior. A given supplier’s actions are less geared towards increasing quantity and more geared towards maintaining input quality after the supplier is integrated and loses access to alternative pay-per-kilo buyers. Finally, we show that firms and individual plants that use integrated suppliers at the time of production ultimately produce a significantly higher share of high quality output. In sum, our results suggest that firms change their organizational structure when their output quality objectives change because controlling the incentives of independent suppliers facing a quantity-quality trade-off is difficult, as classical theories of the firm predict.