WP 522- Reservations and the Politics of Fear
Reserving political office for members of a particular, usually traditionally disadvantaged, group is a common form of political quota in many parts of the world. We first show that if political dysfunction is due to “politics of fear” factors that make individuals support members of their own group for political office (despite their incompetence or venality), then such quotas may improve governance. But this does not occur if the reserved groups are either extremely powerful, or extremely weak. We measure governance outcomes and group strength in rural Indian villages. We show that political office reservations in these villages do indeed improve governance, but only when allocated to groups that are neither too strong or too weak.