WP 521- Time vs. State in Insurance: Experimental Evidence from Contract Farming in Kenya
The gains from insurance arise from the transfer of income across states. Yet, by requiring that the premium be paid upfront, standard insurance products also transfer income across time. We show that this intertemporal transfer can help explain low insurance demand, especially among the poor, and in a randomized control trial in Kenya we test a crop insurance product which removes it. The product is interlinked with a contract farming scheme: as with other inputs, the buyer of the crop offers the insurance and deducts the premium from farmer revenues at harvest time. The take-up rate for pay-at-harvest insurance is 72%, compared to 5% for the standard pay-upfront contract, and take-up is highest among poorer farmers. As for channels, we use additional experiments and outcomes to look at the role of liquidity constraints, present bias, and counterparty risk. Finally, evidence from a natural experiment in the United States, exploiting a change in the timing of the premium payment for Federal Crop Insurance, shows that the transfer across time also affects insurance adoption in developed countries.