WP 516- Digital Credit: A Snapshot of the Current Landscape and Open Research Questions
Lack of access to finance is suspected to be an important impediment to development in low-income countries. A plausible reason that microcredit has been disappointing is that the existing set of credit products might not be appropriate for target customers. In the past few years, digital credit has emerged as an alternative mechanism for providing short-term loans. We summarize the current state of digital credit, focusing primarily on the currently dominant form of credit – consumer loans offered through mobile money systems, often backed by a financial institution. We summarize the current landscape, discuss various ways in which digital credit will represent a change from previously available forms of credit, in particular microcredit or bank loans, and note possible directions for further research.