WP 514- Money and Politics: The Effects of Campaign Spending Limits on Political Competition and Incumbency Advantage
This paper examines the effects of campaign spending limits on political competition and incumbency advantage. We study a reform in Brazil that imposed limits on campaign spending for mayoral elections. These limits were implemented with a discontinuous kink which we exploit for causal identification. We find that stricter limits increase political competition by creating a larger pool of candidates that is on average less wealthy. Moreover, we find that stricter spending limits reduce the incumbency advantage, causing mayors to be less likely to be reelected. These findings are consistent with a contest model with spending caps and endogenous candidate entry.