WP 513- The Political Economy of Program Enforcement: Evidence from Brazil
Do politicians manipulate the enforcement of conditional welfare programs to influence electoral outcomes? We study the Bolsa Familia Program (BFP) in Brazil, which provides a monthly stipend to poor families conditional on school attendance. Repeated failure to comply with this requirement results in increasing penalties. First, we exploit random variation in the timing when beneficiaries learn about penalties for noncompliance around the 2008 municipal elections. We find that the vote share of candidates aligned with the President is lower in zip codes where more beneficiaries received penalties shortly before (as opposed to shortly after) the elections. Second, we show that politicians strategically manipulate enforcement. Using a regression discontinuity design, we find weaker enforcement before elections in municipalities where mayors from the presidential coalition can run for reelection. Finally, we provide evidence that manipulation occurs through misreporting school attendance, particularly in municipalities with a higher fraction of students in schools with politically connected principals.