WP 512- Learning about the Enforcement of Conditional Welfare Programs: Evidence from Brazil
We study the implementation of Bolsa Familia, a program that conditions cash transfers to poor families on children’s school attendance. Using unique administrative data, we analyze how beneficiaries respond to the enforcement of conditionality. Making use of random variation in the day on which punishments are received, we find that school attendance increases after families are punished for past noncompliance. Families also respond to penalties experienced by peers: a child’s attendance increases if her own classmates, but also her siblings’ classmates (in other grades or schools), experience enforcement. As the severity of penalties increases with repeated noncompliance, households’ response is larger when peers receive a penalty that the family has not (yet) received. We thus find evidence of spillover effects and learning about enforcement.