WP 511- Mexicans in America
Using data from the Mexican Family Life Survey (MxFLS), we establish the feasibility and value-added to science of tracking international migrants in a population-representative longitudinal survey. The MxFLS baseline, conducted in 2002, is representative of all Mexicans living in Mexico. The follow-ups are designed to track and interview movers including those who moved to the U.S. In the 2005 follow-up, 91% of baseline respondents who moved to the U.S. were interviewed. Information collected from interviewing these migrants in the U.S. is used to provide scientific evidence on the extent and selectivity of under-enumeration of Mexican-origin migrants in the American Community Survey and Current Population Survey. Between 20 and 35 percent of Mexican-origin migrants who moved between the baseline and follow-up are not enumerated in the U.S.-based surveys. They are more likely to be younger, single, male, and less educated than those who are enumerated in the U.S.-based surveys. 85 percent of MxFLS migrants report entering the U.S. without documentation: they are more likely to be younger, single, male and less educated than those who entered with documentation: the undercount of those who entered without documentation is likely to be larger than the overall estimate. Conditional on age and education, those not enumerated in the U.S.-based surveys are more likely to be working and they earn less than those who were enumerated. Taken together, these results have implications for interpretation of evidence on the number of migrants and their contributions to the U.S. economy and society.