WP 510- Individuals and Organizations as Sources of State Effectiveness, and Consequences for Policy Design
How much of the variation in state effectiveness is due to the individuals and organizations responsible for implementing policy? We investigate this question and its implications for policy design in the context of public procurement, using a text-based product classification method to measure bureaucratic output. We show that effective procurers lower bid preparation/submission costs, and that 60% of within-product purchase-price variation across 16 million purchases in Russia in 2011-2015 is due to the bureaucrats and organizations administering procurement processes. This has dramatic policy consequences. To illustrate these, we study a ubiquitous procurement policy: bid preferences for favored firms (here domestic manufacturers). The policy decreases overall entry and increases prices when procurers are effective, but has the opposite impact with ineffective procurers, as predicted by a simple endogenous-entry model of procurement. Our results imply that the state’s often overlooked bureaucratic tier is critical for effectiveness and the make-up of optimal policies.