WP 508- Competition in Agricultural Markets: An Experimental Approach
This paper presents an experimental approach to measure competition in agricultural markets, based on the random allocation of subsidies to competing traders. We compare prices of subsidized and unsubsidized crop traders to recover the key market structure parameter in a standard model of imperfect competition. By combining the experimental results with quasi-experimental estimates of the pass-through rate, we also estimate market size, or the effective number of traders competing for farmers’ supply. In the context of the Sierra Leone cocoa industry, our results point to a competitive agricultural trading sector and suggest that the market size is substantially larger than the village. The methodology developed in this paper uses purely individual-level treatment to shed light on market structure. This approach may be useful for the many cases in which market-level randomization is not feasible.