WP 504- The Economist as Plumber
As economists increasingly help governments design new policies and regulations, they take on an added responsibility to engage with the details of policy making and, in doing so, to adopt the mindset of a plumber. Plumbers try to predict as well as possible what may work in the real world, mindful that tinkering and adjusting will be necessary since our models gives us very little theoretical guidance on what (and how) details will matter. This essay argues that economists should seriously engage with plumbing, in the interest of both society and our discipline.