WP 502- Adult Mortality Five Year after a Natural Disaster
Exposure to extreme events has been hypothesized to affect subsequent mortality because of mortality selection and scarring effects of the event itself. We examine survival at and in the five years after the 2004 Indian Ocean earthquake and tsunami for a population-representative sample of residents of Aceh, Indonesia who were differentially exposed to the disaster. For this population, the dynamics of selection and scarring are a complex function of the degree of tsunami impact in the community, the nature of individual exposures, age at exposure, and gender. Among individuals from tsunami-affected communities we find evidence for positive mortality selection among older individuals, with stronger effects for males than for females, and that this selection dominates any scarring impact of stressful exposures that elevate mortality. Among individuals from other communities, where mortality selection does not play a role, there is evidence of scarring with property loss associated with elevated mortality risks in the five years after the disaster among adults age 50 or older at the time of the disaster.