WP 501- Impact of Violent Crime on Risk Aversion: Evidence from the Mexican Drug War
Whereas attitudes towards risk are thought to play an important role in many decisions over the life-course, factors that affect those attitudes are not fully understood. Using longitudinal survey data collected in Mexico before and during the Mexican war on drugs, we investigate how an individual’s risk attitudes change with variation in levels of insecurity and uncertainty brought on by unprecedented changes in local-area violent crime due to the war on drugs. Exploiting the fact that the timing, virulence and spatial distribution of changes in violent crime were unanticipated, we establish the changes can plausibly be treated as exogenous in models that also take into account unobserved characteristics of individuals that are fixed over time. As local-area violent crime increases, there is a rise in risk aversion that is distributed through the entire local population.