WP 500- Disrupting Education? Experimental Evidence on Technology-Aided Instruction in India
We present experimental evidence on the impact of a technology-aided after-school instruction program on learning outcomes in middle school grades in urban India, using a lottery that provided students with a voucher to cover program costs. A key feature of the program was its ability to individually customize educational content to match the level and rate of progress of each student. We find that lottery winners had large increases in test scores of 0.36σ in math and 0.22σ in Hindi over just a 4.5-month period. IV estimates suggest that attending the program for 90 days would increase math and Hindi test scores by 0.59σ and 0.36σ respectively. We find similar absolute test score gains for all students, but the relative gain was much greater for academically-weaker students because their rate of learning in the control group was close to zero. We show that the program was able to effectively cater to the very wide variation in student learning levels within a single grade by precisely targeting instruction to the level of student preparation. The program was highly cost-effective, both in terms of productivity per dollar and unit of time. Our results suggest that well-designed technology-aided instruction programs can sharply improve productivity in delivering education.