WP 637 - What Do Market-Access Subsidies Do? Experimental Evidence from Tunisia
"Many countries seek to promote exports by subsidizing market access, but evidence on
such efforts has been mixed. We present the first randomized evaluation of a
government financial-support program explicitly targeting exports, the Tasdir+ program
in Tunisia. The program offered matching grants for fixed market-access costs but not
variable costs. Tracking outcomes in administrative data, we find positive effects on
exports on average. We find limited impacts on the number of destinations or exported
products, which were stated policy targets. The finding that the fixed-cost subsidies
expanded exports on the intensive margin but not the extensive margins of destinations
or products stands in contrast to the predictions of several workhorse trade models."